# docs/spec/openapi.yml
openapi: 3.0.0
info:
  title: Planning Poker API
  version: v1
paths:
  /api/planning-poker/ws:
    get:
      summary: リアルタイム更新のための WebSocket エンドポイント
      description: |
        プランニングポーカーセッションに関するリアルタイム更新を受信するための WebSocket 接続を確立します。
        サーバーは以下のイベントを含むメッセージを送信します:
        - participantJoined: 新しい参加者がセッションに参加しました。
        - roundStarted: 新しいラウンドが開始されました。
        - voteSubmitted: 参加者が投票を送信しました。
        - votesRevealed: ラウンドの投票が公開されました。
        - sessionEnded: セッションが終了しました。
      responses:
        "101":
          description: プロトコルの切り替え
          headers:
            Upgrade:
              schema:
                type: string
              description: websocket
            Connection:
              schema:
                type: string
              description: Upgrade
        "400":
          description: 不正なリクエスト
        "500":
          description: サーバー内部エラー
  /api/planning-poker/sessions:
    post:
      summary: セッションを作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: セッション作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/sessions/{sessionId}/participants:
    post:
      summary: セッションに参加する
      parameters:
        - in: path
          name: sessionId
          schema:
            type: string
            format: uuid
          required: true
          description: セッションID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinSessionRequest"
      responses:
        "201":
          description: セッション参加成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinSessionResponse"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: セッションが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/sessions/{sessionId}:
    get:
      summary: セッションを取得する
      parameters:
        - in: path
          name: sessionId
          schema:
            type: string
            format: uuid
          required: true
          description: セッションID
      responses:
        "200":
          description: セッション取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSessionResponse"
        "404":
          description: セッションが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/sessions/{sessionId}/end:
    post:
      summary: セッションを終了する
      parameters:
        - in: path
          name: sessionId
          schema:
            type: string
            format: uuid
          required: true
          description: セッションID
      responses:
        "200":
          description: セッション終了成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndSessionResponse"
        "404":
          description: セッションが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/sessions/{sessionId}/rounds:
    post:
      summary: ラウンドを開始する
      parameters:
        - in: path
          name: sessionId
          schema:
            type: string
            format: uuid
          required: true
          description: セッションID
      responses:
        "201":
          description: ラウンド開始成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartRoundResponse"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: セッションが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/rounds/{roundId}:
    get:
      summary: ラウンド情報を取得する
      parameters:
        - in: path
          name: roundId
          schema:
            type: string
            format: uuid
          required: true
          description: ラウンドID
      responses:
        "200":
          description: ラウンド情報取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRoundResponse"
        "404":
          description: ラウンドが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/rounds/{roundId}/reveal:
    post:
      summary: ラウンドを終了する
      parameters:
        - in: path
          name: roundId
          schema:
            type: string
            format: uuid
          required: true
          description: ラウンドID
      responses:
        "200":
          description: ラウンド終了成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevealRoundResponse"
        "404":
          description: ラウンドが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/planning-poker/rounds/{roundId}/votes:
    post:
      summary: 投票を送信する
      parameters:
        - in: path
          name: roundId
          schema:
            type: string
            format: uuid
          required: true
          description: ラウンドID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendVoteRequest"
      responses:
        "201":
          description: 投票送信成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendVoteResponse"
        "400":
          description: 不正なリクエスト
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: ラウンドが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      description: エラーレスポンス
      required:
        - message
      properties:
        message:
          type: string
          description: エラーメッセージ
    ScaleType:
      type: string
      description: スケールの種類
      enum:
        - fibonacci
        - t-shirt
        - power-of-two
        - custom
    CreateSessionRequest:
      type: object
      description: セッション作成リクエスト
      required:
        - sessionName
        - scaleType
        - hostName
      properties:
        sessionName:
          type: string
          description: セッション名
        scaleType:
          $ref: "#/components/schemas/ScaleType"
          description: スケールの種類
        customScale:
          type: array
          items:
            type: string
          description: カスタムスケール（scaleTypeがcustomの場合のみ有効）
        hostName:
          type: string
          description: ホスト名
    CreateSessionResponse:
      type: object
      description: セッション作成レスポンス
      required:
        - sessionId
        - hostId
      properties:
        sessionId:
          type: string
          format: uuid
          description: 作成されたセッションのID
        hostId:
          type: string
          format: uuid
          description: セッションのホストID
    JoinSessionRequest:
      type: object
      description: セッション参加リクエスト
      required:
        - name
      properties:
        name:
          type: string
          description: 参加者の名前
    JoinSessionResponse:
      type: object
      description: セッション参加レスポンス
      required:
        - participantId
      properties:
        participantId:
          type: string
          format: uuid
          description: 参加者のID
    GetSessionResponse:
      type: object
      description: セッション取得レスポンス
      required:
        - session
      properties:
        session:
          $ref: "#/components/schemas/Session"
          description: セッション情報
    EndSessionResponse:
      type: object
      description: セッション終了レスポンス
    StartRoundResponse:
      type: object
      description: ラウンド開始レスポンス
      required:
        - roundId
      properties:
        roundId:
          type: string
          format: uuid
          description: 開始されたラウンドのID
    GetRoundResponse:
      type: object
      description: ラウンド情報取得レスポンス
      required:
        - round
      properties:
        round:
          $ref: "#/components/schemas/Round"
          description: ラウンド情報
    RevealRoundResponse:
      type: object
      description: ラウンド結果公開レスポンス
    SendVoteRequest:
      type: object
      description: 投票送信リクエスト
      required:
        - participantId
        - value
      properties:
        participantId:
          type: string
          format: uuid
          description: 投票者のID
        value:
          type: string
          description: 投票値
    SendVoteResponse:
      type: object
      description: 投票送信レスポンス
      required:
        - voteId
      properties:
        voteId:
          type: string
          format: uuid
          description: 投票ID
    SessionParticipant:
      type: object
      description: セッション参加者
      required:
        - participantId
        - name
      properties:
        participantId:
          type: string
          description: 参加者のID
        name:
          type: string
          description: 参加者の名前
    Session:
      type: object
      description: セッション情報
      required:
        - sessionId
        - sessionName
        - hostId
        - scaleType
        - customScale
        - status
        - participants
        - createdAt
        - updatedAt
      properties:
        sessionId:
          type: string
          format: uuid
          description: セッションのID
        sessionName:
          type: string
          description: セッション名
        hostId:
          type: string
          format: uuid
          description: セッションのホストID
        scaleType:
          $ref: "#/components/schemas/ScaleType"
          description: スケールの種類
        customScale:
          type: array
          items:
            type: string
          description: カスタムスケール（scaleTypeがcustomの場合のみ有効）
        currentRoundId:
          type: string
          format: uuid
          description: 現在のラウンドID
        status:
          type: string
          description: セッションの状態
        participants:
          type: array
          items:
            $ref: "#/components/schemas/SessionParticipant"
          description: 参加者リスト
        createdAt:
          type: string
          format: date-time
          description: セッションの作成日時
        updatedAt:
          type: string
          format: date-time
          description: セッションの最終更新日時
    Round:
      type: object
      description: ラウンド情報
      required:
        - roundId
        - sessionId
        - status
        - votes
        - createdAt
        - updatedAt
      properties:
        roundId:
          type: string
          format: uuid
          description: ラウンドのID
        sessionId:
          type: string
          format: uuid
          description: このラウンドが属するセッションのID
        status:
          type: string
          description: ラウンドの状態
          enum:
            - voting
            - revealed
        votes:
          type: array
          items:
            $ref: "#/components/schemas/Vote"
          description: 投票結果のリスト
        createdAt:
          type: string
          format: date-time
          description: ラウンドの作成日時
        updatedAt:
          type: string
          format: date-time
          description: ラウンドの最終更新日時
    Vote:
      type: object
      description: 投票情報
      properties:
        participantId:
          type: string
          format: uuid
          description: 参加者のID
        value:
          type: string
          description: 投票値（ラウンドのstatusがrevealedの場合のみ）
      required:
        - participantId
    WebSocketMessage:
      type: object
      description: WebSocketメッセージ
      properties:
        event:
          type: string
          description: イベント名
          enum:
            - participantJoined
            - roundStarted
            - voteSubmitted
            - votesRevealed
            - sessionEnded
        payload:
          type: object
          description: イベントのペイロード
          oneOf:
            - $ref: "#/components/schemas/ParticipantJoinedPayload"
            - $ref: "#/components/schemas/RoundStartedPayload"
            - $ref: "#/components/schemas/VoteSubmittedPayload"
            - $ref: "#/components/schemas/VotesRevealedPayload"
            - $ref: "#/components/schemas/SessionEndedPayload"
    ParticipantJoinedPayload:
      type: object
      description: participantJoinedイベントのペイロード
      properties:
        participantId:
          type: string
          format: uuid
          description: 参加者のID
        name:
          type: string
          description: 参加者の名前
    RoundStartedPayload:
      type: object
      description: roundStartedイベントのペイロード
      properties:
        roundId:
          type: string
          format: uuid
          description: ラウンドのID
    VoteSubmittedPayload:
      type: object
      description: voteSubmittedイベントのペイロード
      properties:
        participantId:
          type: string
          format: uuid
          description: 投票した参加者のID
    VotesRevealedPayload:
      type: object
      description: votesRevealedイベントのペイロード
      properties:
        votes:
          type: array
          items:
            $ref: "#/components/schemas/Vote"
          description: 投票結果のリスト
        average:
          type: number
          format: float
          description: 投票値の平均
        median:
          type: number
          format: float
          description: 投票値の中央値
    SessionEndedPayload:
      type: object
      description: sessionEndedイベントのペイロード